refactor: add `StudyDeploymentProxy` to remove coupling of `StudyRuntime` to services
Tests in `StudyRuntimeTest` have been moved to `StudyDeploymentProxyTest`. But, `StudyRuntime` with less dependencies can now be introduced, and maybe some `StudyDeploymentProxyTest` will become superfluous.
