+++ carp.clients.core/src/commonTest/kotlin/dk/cachet/carp/clients/domain/ClientRepositoryTest.kt
@@ -1,14 +1,8 @@
 package dk.cachet.carp.clients.domain
 
-import dk.cachet.carp.clients.domain.data.DataListener
 import dk.cachet.carp.common.application.UUID
-import dk.cachet.carp.common.application.devices.SmartphoneDeviceRegistration
-import dk.cachet.carp.common.application.services.createApplicationServiceAdapter
-import dk.cachet.carp.common.infrastructure.services.SingleThreadedEventBus
-import dk.cachet.carp.data.infrastructure.InMemoryDataStreamService
-import dk.cachet.carp.deployments.application.DeploymentService
-import dk.cachet.carp.deployments.application.DeploymentServiceHost
-import dk.cachet.carp.deployments.infrastructure.InMemoryDeploymentRepository
+import dk.cachet.carp.common.infrastructure.test.StubMasterDeviceDescriptor
+import dk.cachet.carp.deployments.application.MasterDeviceDeployment
 import dk.cachet.carp.test.runSuspendTest
 import kotlin.test.*
 
@@ -20,71 +14,44 @@ interface ClientRepositoryTest
      */
     fun createRepository(): ClientRepository
 
-    private fun createDependencies(): Triple<ClientRepository, DeploymentService, DataListener>
-    {
-        val eventBus = SingleThreadedEventBus()
-
-        val deploymentService = DeploymentServiceHost(
-            InMemoryDeploymentRepository(),
-            InMemoryDataStreamService(),
-            eventBus.createApplicationServiceAdapter( DeploymentService::class ) )
-        return Triple( createRepository(), deploymentService, createDataListener() )
-    }
-
-    private suspend fun addTestDeployment( deploymentService: DeploymentService ): UUID
-    {
-        val protocol = createSmartphoneStudy()
-        val invitation = createParticipantInvitation( protocol )
-        val status = deploymentService.createStudyDeployment( UUID.randomUUID(), protocol.getSnapshot(), listOf( invitation ) )
-
-        return status.studyDeploymentId
-    }
-
-    private suspend fun createTestStudyRuntime(): Pair<ClientRepository, StudyRuntime>
-    {
-        val (repo, deploymentService, dataListener) = createDependencies()
-        val deploymentId = addTestDeployment( deploymentService )
-        val roleName = smartphone.roleName
-        return repo to StudyRuntime.initialize(
-            deploymentService, dataListener,
-            deploymentId, roleName, smartphone.createRegistration() )
-    }
-
 
     @Test
     fun deviceRegistration_is_initially_null() = runSuspendTest {
-        val (repo, _) = createDependencies()
+        val repo = createRepository()
         assertNull( repo.getDeviceRegistration() )
     }
 
     @Test
     fun addStudyRuntime_can_be_retrieved() = runSuspendTest {
-        val (repo, studyRuntime) = createTestStudyRuntime()
-        repo.addStudyRuntime( studyRuntime )
+        val repo = createRepository()
+
+        val runtime = StudyRuntime( UUID.randomUUID(), "Device role" )
+        repo.addStudyRuntime( runtime )
 
         // Runtime can be retrieved by ID.
-        val deploymentId = studyRuntime.studyDeploymentId
-        val roleName = studyRuntime.device.roleName
+        val deploymentId = runtime.studyDeploymentId
+        val roleName = runtime.deviceRoleName
         val retrievedRuntime = repo.getStudyRuntimeBy( deploymentId, roleName )
         assertNotNull( retrievedRuntime )
 
         // Runtime is included in list.
         val allRuntimes = repo.getStudyRuntimeList()
         assertEquals( 1, allRuntimes.count() )
-        assertNotNull( allRuntimes.single { it.studyDeploymentId == deploymentId && it.device.roleName == roleName } )
+        assertNotNull( allRuntimes.single { it.studyDeploymentId == deploymentId && it.deviceRoleName == roleName } )
     }
 
     @Test
     fun addStudyRuntime_fails_for_existing_runtime() = runSuspendTest {
-        val (repo, studyRuntime) = createTestStudyRuntime()
-        repo.addStudyRuntime( studyRuntime )
+        val repo = createRepository()
+        val runtime = StudyRuntime( UUID.randomUUID(), "Device role" )
+        repo.addStudyRuntime( runtime )
 
-        assertFailsWith<IllegalArgumentException> { repo.addStudyRuntime( studyRuntime ) }
+        assertFailsWith<IllegalArgumentException> { repo.addStudyRuntime( runtime ) }
     }
 
     @Test
     fun getStudyRuntimeBy_is_null_for_unknown_runtime() = runSuspendTest {
-        val (repo, _) = createDependencies()
+        val repo = createRepository()
 
         val unknownId = UUID.randomUUID()
         assertNull( repo.getStudyRuntimeBy( unknownId, "Unknown" ) )
@@ -92,57 +59,59 @@ interface ClientRepositoryTest
 
     @Test
     fun getStudyRuntimeList_is_empty_initially() = runSuspendTest {
-        val (repo, _) = createDependencies()
+        val repo = createRepository()
 
         assertEquals( 0, repo.getStudyRuntimeList().count() )
     }
 
     @Test
     fun updateStudyRuntime_succeeds() = runSuspendTest {
-        val (repo, deploymentService, dataListener) = createDependencies()
-        val protocol = createDependentSmartphoneStudy()
-        val invitation = createParticipantInvitation( protocol )
+        val repo = createRepository()
         val deploymentId = UUID.randomUUID()
-        deploymentService.createStudyDeployment( deploymentId, protocol.getSnapshot(), listOf( invitation ) )
-        val studyRuntime = StudyRuntime.initialize(
-            deploymentService, dataListener,
-            deploymentId, smartphone.roleName, smartphone.createRegistration() )
-        repo.addStudyRuntime( studyRuntime )
+        val deviceRoleName = "Device role"
+        val runtime = StudyRuntime( deploymentId, deviceRoleName )
+        repo.addStudyRuntime( runtime )
 
         // Make some changes and update.
-        deploymentService.registerDevice( deploymentId, deviceSmartphoneDependsOn.roleName, SmartphoneDeviceRegistration( "dependent" ) )
-        studyRuntime.tryDeployment( deploymentService, dataListener )
-        repo.updateStudyRuntime( studyRuntime )
+        val masterDevice = StubMasterDeviceDescriptor( deviceRoleName )
+        val registration = masterDevice.createRegistration()
+        val masterDeviceDeployment = MasterDeviceDeployment( StubMasterDeviceDescriptor( deviceRoleName ), registration )
+        runtime.deploymentReceived( masterDeviceDeployment, emptySet() )
+        runtime.completeDeployment( createDataListener() )
+        repo.updateStudyRuntime( runtime )
 
         // Verify whether changes were stored.
-        val retrievedRuntime = repo.getStudyRuntimeBy( deploymentId, smartphone.roleName )
+        val retrievedRuntime = repo.getStudyRuntimeBy( deploymentId, deviceRoleName )
         assertNotNull( retrievedRuntime )
-        assertEquals( studyRuntime.getSnapshot(), retrievedRuntime.getSnapshot() )
+        assertEquals( runtime.getSnapshot(), retrievedRuntime.getSnapshot() )
     }
 
     @Test
     fun updateStudyRuntime_fails_for_unknown_runtime() = runSuspendTest {
-        val (repo, studyRuntime) = createTestStudyRuntime()
+        val repo = createRepository()
 
-        assertFailsWith<IllegalArgumentException> { repo.updateStudyRuntime( studyRuntime ) }
+        val runtime = StudyRuntime( UUID.randomUUID(), "Device role" )
+        assertFailsWith<IllegalArgumentException> { repo.updateStudyRuntime( runtime ) }
     }
 
     @Test
     fun removeStudyRuntime_succeeds() = runSuspendTest {
-        val (repo, studyRuntime) = createTestStudyRuntime()
-        repo.addStudyRuntime( studyRuntime )
+        val repo = createRepository()
+        val runtime = StudyRuntime( UUID.randomUUID(), "Device role" )
+        repo.addStudyRuntime( runtime )
 
-        repo.removeStudyRuntime( studyRuntime )
+        repo.removeStudyRuntime( runtime )
 
-        val deploymentId = studyRuntime.studyDeploymentId
-        val roleName = studyRuntime.device.roleName
+        val deploymentId = runtime.studyDeploymentId
+        val roleName = runtime.deviceRoleName
         assertNull( repo.getStudyRuntimeBy( deploymentId, roleName ) )
         assertEquals( 0, repo.getStudyRuntimeList().count() )
     }
 
     @Test
     fun removeStudyRuntime_succeeds_when_runtime_not_present() = runSuspendTest {
-        val (repo, studyRuntime) = createTestStudyRuntime()
-        repo.removeStudyRuntime( studyRuntime )
+        val repo = createRepository()
+        val runtime = StudyRuntime( UUID.randomUUID(), "Device role" )
+        repo.removeStudyRuntime( runtime )
     }
 }
