+++ carp.clients.core/src/commonTest/kotlin/dk/cachet/carp/clients/domain/study/StudyDeploymentProxyTest.kt
@@ -1,0 +1,351 @@
+package dk.cachet.carp.clients.domain.study
+
+import dk.cachet.carp.clients.domain.connectedDevice
+import dk.cachet.carp.clients.domain.createDataListener
+import dk.cachet.carp.clients.domain.createDependentSmartphoneStudy
+import dk.cachet.carp.clients.domain.createSmartphoneStudy
+import dk.cachet.carp.clients.domain.createSmartphoneWithConnectedDeviceStudy
+import dk.cachet.carp.clients.domain.createStudyDeployment
+import dk.cachet.carp.clients.domain.deviceSmartphoneDependsOn
+import dk.cachet.carp.clients.domain.smartphone
+import dk.cachet.carp.clients.domain.DeviceRegistrationStatus
+import dk.cachet.carp.clients.domain.StudyRuntime
+import dk.cachet.carp.clients.domain.StudyRuntimeStatus
+import dk.cachet.carp.clients.domain.data.AnyConnectedDeviceDataCollector
+import dk.cachet.carp.clients.domain.data.DataListener
+import dk.cachet.carp.clients.domain.data.DeviceDataCollectorFactory
+import dk.cachet.carp.clients.domain.data.StubConnectedDeviceDataCollectorFactory
+import dk.cachet.carp.clients.domain.data.StubDeviceDataCollector
+import dk.cachet.carp.common.application.UUID
+import dk.cachet.carp.common.application.data.DataType
+import dk.cachet.carp.common.application.devices.AltBeaconDeviceRegistration
+import dk.cachet.carp.common.application.devices.DeviceRegistration
+import dk.cachet.carp.common.application.devices.DeviceType
+import dk.cachet.carp.common.application.tasks.Measure
+import dk.cachet.carp.common.infrastructure.test.STUB_DATA_TYPE
+import dk.cachet.carp.common.infrastructure.test.StubDeviceDescriptor
+import dk.cachet.carp.common.infrastructure.test.StubTaskDescriptor
+import dk.cachet.carp.deployments.application.DeviceDeploymentStatus
+import dk.cachet.carp.deployments.application.StudyDeploymentStatus
+import dk.cachet.carp.protocols.domain.start
+import dk.cachet.carp.test.runSuspendTest
+import kotlin.test.*
+
+
+/**
+ * Tests for [StudyDeploymentProxy].
+ */
+class StudyDeploymentProxyTest
+{
+    @Test
+    fun tryDeployment_deploys_when_possible() = runSuspendTest {
+        // Create a deployment service which contains a 'smartphone study'.
+        val (deploymentService, deploymentStatus) = createStudyDeployment( createSmartphoneStudy() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+
+        // Study runtime status is deployed and contains registered master device.
+        assertTrue( runtime.isDeployed )
+        val runtimeStatus = runtime.getStatus()
+        assertTrue( runtimeStatus is StudyRuntimeStatus.Deployed )
+        val registrationStatus = runtimeStatus.devicesRegistrationStatus.values.singleOrNull()
+        assertTrue( registrationStatus is DeviceRegistrationStatus.Registered )
+        assertEquals( smartphone, registrationStatus.device )
+        assertEquals( deviceRegistration, registrationStatus.registration )
+
+        // Study runtime events reflects deployment has been received and completed.
+        val events = runtime.consumeEvents()
+        val receivedEvent = events.filterIsInstance<StudyRuntime.Event.DeploymentReceived>()
+        assertEquals( 1, receivedEvent.count() )
+        assertEquals( runtimeStatus.deploymentInformation, receivedEvent.single().deploymentInformation )
+        assertEquals( 1, events.filterIsInstance<StudyRuntime.Event.DeploymentCompleted>().count() )
+
+        // Master device status in deployment is also set to deployed.
+        val newDeploymentStatus = deploymentService.getStudyDeploymentStatus( deploymentStatus.studyDeploymentId )
+        val masterDeviceStatus = newDeploymentStatus.getDeviceStatus( smartphone )
+        assertTrue( masterDeviceStatus is DeviceDeploymentStatus.Deployed )
+    }
+
+    @Test
+    fun tryDeployment_does_not_deploy_when_depending_on_other_devices() = runSuspendTest {
+        // Create a deployment service which contains a study where 'smartphone' depends on another master device.
+        val (deploymentService, deploymentStatus) = createStudyDeployment( createDependentSmartphoneStudy() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+
+        // Study runtime status is not ready for deployment.
+        assertFalse( runtime.isDeployed )
+        val runtimeStatus = runtime.getStatus()
+        assertTrue( runtimeStatus is StudyRuntimeStatus.NotReadyForDeployment )
+
+        // Study runtime events reflects deployment has not been received yet.
+        val events = runtime.consumeEvents()
+        assertEquals( 0, events.filterIsInstance<StudyRuntime.Event.DeploymentReceived>().count() )
+
+        // Master device status in deployment is registered, but not deployed.
+        val newDeploymentStatus = deploymentService.getStudyDeploymentStatus( deploymentStatus.studyDeploymentId )
+        val masterDeviceStatus = newDeploymentStatus.getDeviceStatus( smartphone )
+        assertTrue( masterDeviceStatus is DeviceDeploymentStatus.Registered )
+    }
+
+    @Test
+    fun tryDeployment_does_not_deploy_when_registering_devices() = runSuspendTest {
+        // Create a deployment service which contains a study where 'smartphone' depends on a connected device.
+        val (deploymentService, deploymentStatus) =
+            createStudyDeployment( createSmartphoneWithConnectedDeviceStudy() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+
+        // Study runtime status indicates `connectedDevice` needs to be registered.
+        assertFalse( runtime.isDeployed )
+        val runtimeStatus = runtime.getStatus()
+        assertTrue( runtimeStatus is StudyRuntimeStatus.RegisteringDevices )
+        assertEquals( connectedDevice, runtimeStatus.remainingDevicesToRegister.single() )
+        val connectedRegistrationStatus = runtimeStatus.devicesRegistrationStatus[ connectedDevice ]
+        assertTrue( connectedRegistrationStatus is DeviceRegistrationStatus.Unregistered )
+        assertEquals( connectedDevice, connectedRegistrationStatus.device )
+
+        // Study runtime events reflects deployment has been received, but not completed.
+        val events = runtime.consumeEvents()
+        val receivedEvent = events.filterIsInstance<StudyRuntime.Event.DeploymentReceived>()
+        assertEquals( 1, receivedEvent.count() )
+        assertEquals( runtimeStatus.deploymentInformation, receivedEvent.single().deploymentInformation )
+        assertEquals( 0, events.filterIsInstance<StudyRuntime.Event.DeploymentCompleted>().count() )
+
+        // Master device status in deployment is registered, but not deployed.
+        val newDeploymentStatus = deploymentService.getStudyDeploymentStatus( deploymentStatus.studyDeploymentId )
+        val masterDeviceStatus = newDeploymentStatus.getDeviceStatus( smartphone )
+        assertTrue( masterDeviceStatus is DeviceDeploymentStatus.Registered )
+    }
+
+    @Test
+    fun tryDeployment_fails_for_unknown_studyDeploymentId() = runSuspendTest {
+        // Create a deployment service which contains a 'smartphone study'.
+        val (deploymentService, _) = createStudyDeployment( createSmartphoneStudy() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+
+        val unknownId = UUID.randomUUID()
+        val runtime = StudyRuntime( unknownId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        assertFailsWith<IllegalArgumentException> {
+            studyDeployment.tryDeployment( runtime, deviceRegistration )
+        }
+    }
+
+    @Test
+    fun tryDeployment_fails_for_unknown_deviceRoleName() = runSuspendTest {
+        val (deploymentService, deploymentStatus) = createStudyDeployment( createSmartphoneStudy() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, "Unknown role" )
+        val deviceRegistration = smartphone.createRegistration()
+        assertFailsWith<IllegalArgumentException> {
+            studyDeployment.tryDeployment( runtime, deviceRegistration )
+        }
+    }
+
+    @Test
+    fun tryDeployment_fails_for_incorrect_deviceRegistration() = runSuspendTest {
+        val (deploymentService, deploymentStatus) = createStudyDeployment( createSmartphoneStudy() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val incorrectRegistration = AltBeaconDeviceRegistration( 0, UUID.randomUUID(), 0, 0, 0 )
+        assertFailsWith<IllegalArgumentException> {
+            studyDeployment.tryDeployment( runtime, incorrectRegistration )
+        }
+    }
+
+    @Test
+    fun tryDeployment_only_succeeds_after_ready_for_deployment() = runSuspendTest {
+        // Create a study runtime for a study where 'smartphone' depends on another master device ('deviceSmartphoneDependsOn').
+        val (deploymentService, deploymentStatus) = createStudyDeployment( createDependentSmartphoneStudy() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+
+        // Dependent devices are not yet registered.
+        var status = runtime.getStatus()
+        assertEquals( 0, runtime.consumeEvents().filterIsInstance<StudyRuntime.Event.DeploymentCompleted>().count() )
+        assertTrue( status is StudyRuntimeStatus.NotReadyForDeployment )
+
+        // Once dependent devices are registered, deployment succeeds.
+        deploymentService.registerDevice(
+            deploymentStatus.studyDeploymentId,
+            deviceSmartphoneDependsOn.roleName,
+            deviceSmartphoneDependsOn.createRegistration() )
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+        status = runtime.getStatus()
+        assertEquals( 1, runtime.consumeEvents().filterIsInstance<StudyRuntime.Event.DeploymentCompleted>().count() )
+        assertTrue( status is StudyRuntimeStatus.Deployed )
+        val registrationStatus = status.devicesRegistrationStatus.values.singleOrNull()
+        assertTrue( registrationStatus is DeviceRegistrationStatus.Registered )
+        assertEquals( smartphone, registrationStatus.device )
+        assertEquals( deviceRegistration, registrationStatus.registration )
+    }
+
+    @Test
+    fun tryDeployment_only_succeeds_after_devices_are_registered() = runSuspendTest {
+        // Create a study runtime for a study where 'smartphone' depends on a connected device.
+        val (deploymentService, deploymentStatus) =
+            createStudyDeployment( createSmartphoneWithConnectedDeviceStudy() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+
+        // Connected device is not yet registered.
+        var status = runtime.getStatus()
+        assertEquals( 0, runtime.consumeEvents().filterIsInstance<StudyRuntime.Event.DeploymentCompleted>().count() )
+        assertTrue( status is StudyRuntimeStatus.RegisteringDevices )
+
+        // Once device is registered, deployment succeeds.
+        // TODO: It should be possible to register this device through `StudyManager` rather than directly from `deploymentService`.
+        deploymentService.registerDevice(
+            deploymentStatus.studyDeploymentId,
+            connectedDevice.roleName,
+            connectedDevice.createRegistration() )
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+        status = runtime.getStatus()
+        assertEquals( 1, runtime.consumeEvents().filterIsInstance<StudyRuntime.Event.DeploymentCompleted>().count() )
+        assertTrue( status is StudyRuntimeStatus.Deployed )
+        val registrationStatuses = status.devicesRegistrationStatus
+        assertEquals( 2, registrationStatuses.size ) // Smartphone and connected device.
+        assertTrue( registrationStatuses[ smartphone ] is DeviceRegistrationStatus.Registered )
+        assertTrue( registrationStatuses[ connectedDevice ] is DeviceRegistrationStatus.Registered )
+    }
+
+    @Test
+    fun tryDeployment_changes_nothing_when_already_deployed() = runSuspendTest {
+        // Create a study runtime which instantly deploys because the protocol only contains one master device.
+        val (deploymentService, deploymentStatus) = createStudyDeployment( createSmartphoneStudy() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+        assertTrue( runtime.isDeployed )
+
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+        val status = runtime.getStatus()
+        assertTrue( status is StudyRuntimeStatus.Deployed )
+    }
+
+    @Test
+    fun tryDeployment_succeeds_when_data_types_of_protocol_measures_are_supported() = runSuspendTest {
+        // Create protocol that measures on smartphone and one connected device.
+        val protocol = createSmartphoneWithConnectedDeviceStudy()
+        val masterTask = StubTaskDescriptor( "Master measure", listOf( Measure.DataStream( STUB_DATA_TYPE ) ) )
+        protocol.addTaskControl( smartphone.atStartOfStudy().start( masterTask, smartphone ) )
+        val connectedDataType = DataType( "custom", "type" )
+        val connectedTask = StubTaskDescriptor( "Connected measure", listOf( Measure.DataStream( connectedDataType ) ) )
+        protocol.addTaskControl( smartphone.atStartOfStudy().start( connectedTask, connectedDevice ) )
+
+        // Create a data listener which supports the requested devices and types in the protocol
+        val dataListener = DataListener( StubConnectedDeviceDataCollectorFactory(
+            localSupportedDataTypes = setOf( STUB_DATA_TYPE ),
+            mapOf( StubDeviceDescriptor::class to setOf( connectedDataType ) )
+        ) )
+
+        // Create study deployment with preregistered connected device (otherwise study runtime initialization won't complete).
+        val (deploymentService, deploymentStatus) = createStudyDeployment( protocol )
+        deploymentService.registerDevice(
+            deploymentStatus.studyDeploymentId,
+            connectedDevice.roleName,
+            connectedDevice.createRegistration() )
+
+        // Initializing study runtime for the smartphone deployment should succeed since devices and data types are supported.
+        val studyDeployment = StudyDeploymentProxy( deploymentService, dataListener )
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+        assertTrue( runtime.isDeployed )
+    }
+
+    @Test
+    fun tryDeployment_fails_when_requested_data_cannot_be_collected() = runSuspendTest {
+        // Create a protocol that has one measure.
+        val protocol = createSmartphoneStudy()
+        val task = StubTaskDescriptor( "One measure", listOf( Measure.DataStream( STUB_DATA_TYPE ) ) )
+        protocol.addTaskControl( smartphone.atStartOfStudy().start( task, smartphone ) )
+
+        // Initializing study runtime for the smartphone deployment should fail since StubMeasure can't be collected.
+        val (deploymentService, deploymentStatus) = createStudyDeployment( protocol )
+        val dataListener = createDataListener( supportedDataTypes = emptyArray() )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, dataListener )
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        assertFailsWith<UnsupportedOperationException>
+        {
+            studyDeployment.tryDeployment( runtime, deviceRegistration )
+        }
+    }
+
+    @Test
+    fun tryDeployment_fails_when_connected_device_is_not_supported() = runSuspendTest {
+        // Create a deployment for a protocol with a preregistered connected device but no measures.
+        val (deploymentService, deploymentStatus) =
+            createStudyDeployment( createSmartphoneWithConnectedDeviceStudy() )
+        deploymentService.registerDevice(
+            deploymentStatus.studyDeploymentId,
+            connectedDevice.roleName,
+            connectedDevice.createRegistration() )
+
+        // Create a listener which does not support measuring on the connected device.
+        val localDataCollector = StubDeviceDataCollector( emptySet() )
+        val factory =
+            object : DeviceDataCollectorFactory( localDataCollector )
+            {
+                override fun createConnectedDataCollector(
+                    deviceType: DeviceType,
+                    deviceRegistration: DeviceRegistration
+                ): AnyConnectedDeviceDataCollector = throw UnsupportedOperationException( "Unsupported device type." )
+            }
+        val dataListener = DataListener( factory )
+
+        // Even though there are no measures for the connected device in the protocol, it should still verify support.
+        val studyDeployment = StudyDeploymentProxy( deploymentService, dataListener )
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val deviceRegistration = smartphone.createRegistration()
+        assertFailsWith<UnsupportedOperationException>
+        {
+            studyDeployment.tryDeployment( runtime, deviceRegistration )
+        }
+    }
+
+    @Test
+    fun stop_succeeds() = runSuspendTest {
+        // Initialize a study runtime for a typical 'smartphone study'.
+        val (deploymentService, deploymentStatus) = createStudyDeployment( createSmartphoneStudy() )
+        val runtime = StudyRuntime( deploymentStatus.studyDeploymentId, smartphone.roleName )
+        val studyDeployment = StudyDeploymentProxy( deploymentService, createDataListener() )
+        val deviceRegistration = smartphone.createRegistration()
+        studyDeployment.tryDeployment( runtime, deviceRegistration )
+        check( runtime.isDeployed )
+        runtime.consumeEvents() // Drop events so only new ones under test appear.
+
+        studyDeployment.stop( runtime )
+        val status = runtime.getStatus()
+
+        // Study runtime status reflects the study has stopped.
+        assertTrue( runtime.isStopped )
+        assertTrue( status is StudyRuntimeStatus.Stopped )
+        assertTrue( runtime.isDeployed ) // The device is still considered deployed.
+
+        // Study runtime events reflects deployment has stopped
+        assertEquals( 1, runtime.consumeEvents().filterIsInstance<StudyRuntime.Event.DeploymentStopped>().count() )
+
+        // Deployment status also reflects deployment has stopped.
+        val newDeploymentStatus = deploymentService.getStudyDeploymentStatus( deploymentStatus.studyDeploymentId )
+        assertTrue( newDeploymentStatus is StudyDeploymentStatus.Stopped )
+    }
+}
